refactor auth location
