+++ app/util/appUtil.js
@@ -5,8 +5,7 @@ const BASE58 = '123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz'
 const bs58 = require('base-x')(BASE58)
 const fetch = require('node-fetch')
 const FormData = require('form-data')
-const jwksRsa = require('jwks-rsa')
-const jwt = require('jsonwebtoken')
+
 const {
   types: {
     Role: { _enum: rolesEnum },
@@ -20,9 +19,6 @@ const {
   METADATA_KEY_LENGTH,
   METADATA_VALUE_LITERAL_LENGTH,
   MAX_METADATA_COUNT,
-  AUTH_AUDIENCE,
-  AUTH_JWKS_URI,
-  AUTH_ISSUER,
   PROCESS_IDENTIFIER_LENGTH,
 } = require('../env')
 const logger = require('../logger')
@@ -456,49 +452,6 @@ const getMetadataResponse = async (tokenId, metadataKey, res) => {
   return
 }
 
-const client = jwksRsa({
-  cache: true,
-  rateLimit: true,
-  jwksRequestsPerMinute: 5,
-  jwksUri: AUTH_JWKS_URI,
-})
-
-async function getKey(header, cb) {
-  client.getSigningKey(header.kid, (err, key) => {
-    if (err) {
-      logger.warn(`An error occurred getting jwks key ${err}`)
-      cb(err, null)
-    } else if (key) {
-      const signingKey = key.publicKey || key.rsaPublicKey
-      cb(null, signingKey)
-    }
-  })
-}
-
-const verifyJwks = async (authHeader) => {
-  const authToken = authHeader ? authHeader.replace('Bearer ', '') : ''
-
-  const verifyOptions = {
-    audience: AUTH_AUDIENCE,
-    issuer: [AUTH_ISSUER],
-    algorithms: ['RS256'],
-    header: authToken,
-  }
-
-  return new Promise((resolve, reject) => {
-    jwt.verify(authToken, getKey, verifyOptions, (err, decoded) => {
-      if (err) {
-        resolve(false)
-      } else if (decoded) {
-        resolve(true)
-      } else {
-        logger.warn(`Error verifying jwks`)
-        reject({ message: 'An error occurred during jwks verification' })
-      }
-    })
-  })
-}
-
 module.exports = {
   runProcess,
   getMembers,
@@ -517,6 +470,5 @@ module.exports = {
   rolesEnum,
   containsInvalidMembershipRoles,
   getMetadataResponse,
-  verifyJwks,
   validateProcess,
 }
